requirements.txt
aiogram==3.12.0
yt-dlp==2024.08.01
redis==5.0.8
python-dotenv==1.0.1
aiohttp==3.10.5
ffmpeg-python==0.2.0
arq==0.25.0
sqlalchemy[asyncio]==2.0.32
aiosqlite==0.20.0
structlog==24.4.0
